split PostDetail into client part and server part
